+++ lib/msal-browser-poc/src/utils/Constants.ts
@@ -7,7 +7,6 @@
  * @hidden
  * Constants File which contains string constants used by this library
  */
-import {  } from "msal-common";
 
 // String constant to represent interaction status
 export const INTERACTION_IN_PROGRESS = "interaction_in_progress";
