+++ lib/msal-common-poc/src/index.ts
@@ -14,6 +14,8 @@ export { AccessTokenValue } from "./cache/AccessTokenValue";
 export { INetworkModule } from "./app/INetworkModule";
 export { IUri } from "./url/IUri";
 export { UrlString } from "./url/UrlString";
+// Crypto Interface
+export { ICrypto, PKCECodes } from "./utils/crypto/ICrypto";
 // Request and Response
 export { AuthenticationParameters } from "./request/AuthenticationParameters";
 export { AuthResponse, buildResponseStateOnly } from "./response/AuthResponse";
