+++ lib/msal-common-poc/src/utils/TokenUtils.ts
@@ -5,6 +5,7 @@
 
 import { CryptoUtils } from "./CryptoUtils";
 import { StringUtils } from "./StringUtils";
+import { ICrypto } from "./crypto/ICrypto";
 
 /**
  * @hidden
@@ -39,7 +40,7 @@ export class TokenUtils {
      *
      * @param encodedIdToken
      */
-    static extractIdToken(encodedIdToken: string): any {
+    static extractIdToken(encodedIdToken: string, crypto: ICrypto): any {
     // id token will be decoded to get the username
         const decodedToken = this.decodeJwt(encodedIdToken);
         if (!decodedToken) {
@@ -47,7 +48,7 @@ export class TokenUtils {
         }
         try {
             const base64IdToken = decodedToken.JWSPayload;
-            const base64Decoded = CryptoUtils.base64Decode(base64IdToken);
+            const base64Decoded = crypto.base64Decode(base64IdToken);
             if (!base64Decoded) {
                 // this._requestContext.logger.info("The returned id_token could not be base64 url safe decoded.");
                 return null;
